organizacion de todo
