--- #LibroPrueba.py
@@ -1,2 +1,0 @@
-class LibroPrueba:
-    pass
