--- example.dio
@@ -1,1 +1,0 @@
-example
